🐛 Fix main path
When installed via `go install`, the binary is being output as `cmd`
due to using the incorrect path. Using the name of the linter within
this means it gets compiled to the expected name.

Alternatively main could be shifted to the root. This has been
avoided (at least for now) to separate it from the code that would
be used if the linter is to be included in other tooling.