🐛 Fix Eval of Nested Ifs
Evaluation of whether an if returns when it contains a nested
if/else reported false-positives. It did this in situations where
the nested if returned, but the else case(s) did not.

For the case being directly analyzed, only the if is relevant.
However to evaluate that if, any nested statements (including
nested if/else), all cases must be fully checked.