🐛 Fix Nested Assign in Else
Report redundant `else` when a nested if/else uses an
assignment in the `else`.

To achieve this, it is necessary to continue to analyze more
deeply regardless of whether the current node returns. It is
also necessary to understand when part of an if/else, and
whether the earlier parts of that will always return.

The changes also remove the need to keep track of the
depth, as `ifAlwaysReturns` is called directly for the
current node.